Cleanup.
Removed unused Model.serializeGeneric().
Removed unused Model.restoreAttributes().
Changed Store.syncRecord() to populate Model._protectedRelationships in all cases.
Changed the behavior of Store.destroy() when the model is part of a to-one relationship; set the related property to null instead of substituting a dummy Model.
Added some tests around Store.destroy() when the model is referenced by other models in the store.